Abstracting question score calc logic to be re-used by external grader when partials returned
